requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
flask_jwt_extended==4.7.1
charset_normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.2
qrcode==8.2
Pillow==12.0.0
